+++ src/finaletools/cli/main_cli.py
@@ -14,7 +14,7 @@ from finaletools.frag.adjust_wps import adjust_wps
 from finaletools.frag.agg_wps import agg_wps
 from finaletools.frag.delfi_gc_correct import cli_delfi_gc_correct
 from finaletools.frag.end_motifs import (
-    end_motifs, _cli_mds, interval_end_motifs)
+    end_motifs, _cli_mds, _cli_interval_mds, interval_end_motifs)
 from finaletools.frag.cleavage_profile import _cli_cleavage_profile
 from finaletools.genome.gaps import _cli_gap_bed
 
@@ -736,6 +736,7 @@ def main_cli_parser():
     )
     parser_command10a.set_defaults(func=interval_end_motifs)
 
+
     # Subcommand 11: MDS
     parser_command11 = subparsers.add_parser(
         'mds',
@@ -766,7 +767,37 @@ def main_cli_parser():
     )
     parser_command11.set_defaults(func=_cli_mds)
 
+
+    # Subcommand 11a: interval-mds
+    parser_command11a = subparsers.add_parser(
+        'interval-mds',
+        prog='finaletools-interval-mds',
+        description='Reads k-mer frequencies from a file and calculates a '
+        'motif diversity score (MDS) for each interval using normalized '
+        'Shannon entropy as '
+        'described by Jiang et al (2020). This function is generalized for '
+        'any k-mer instead of just 4-mers.'
+    )
+    parser_command11a.add_argument(
+        'file_path',
+        nargs='?',
+        default='-',
+        help='Tab-delimited or similar file containing one column for all '
+        'k-mers a one column for frequency. Reads from stdin by default.'
+    )
+    parser_command11a.add_argument(
+        '-s',
+        '--sep',
+        default=',',
+        help='Separator used in tabular file. Default is tab.'
+    )
+    parser_command11a.add_argument(
+        'file_out',
+        default='-'
+    )
+    parser_command11a.set_defaults(func=_cli_interval_mds)
     
+
     # Subcommand 12: gap bed
     parser_command12 = subparsers.add_parser(
         'gap-bed',